feat(common): select FieldElement by field feature and relocate WhirR1CSSchemeBuilder
- Split the `FieldElement` alias: `ark_bn254::Fr` under `bn254`,
  `whir::algebra::fields::Field64_3` under `goldilocks`. Gate the
  BN254-only modules (mavros, noir_proof_scheme, ntt, optimize, poseidon2,
  prover, skyscraper, verifier) and the Skyscraper/Poseidon2 engine
  registrations in `register_ntt`.
- Move `WhirR1CSSchemeBuilder` (and the `MIN_*` consts) from
  `r1cs-compiler` into `common`, re-exporting it from both crates. The
  move rides here because the crate-root re-export and the field split
  touch the same `lib.rs`. The Mavros constructor stays `bn254`-gated.
