+++ provekit/common/src/lib.rs
@@ -1,20 +1,36 @@
+// `bn254` takes precedence when both field features are enabled (CI builds
+// with `--all-features`, which unavoidably turns both on). A goldilocks
+// build must therefore disable the default:
+//   cargo build -p provekit-common -p provekit-prover -p provekit-verifier \
+//       --no-default-features --features goldilocks
+#[cfg(not(any(feature = "bn254", feature = "goldilocks")))]
+compile_error!("enable a field feature: `bn254` (default) or `goldilocks`");
+
 pub mod file;
 pub use file::binary_format;
 pub mod hash_config;
 mod interner;
+#[cfg(feature = "bn254")]
 mod mavros;
+#[cfg(feature = "bn254")]
 mod noir_proof_scheme;
+#[cfg(feature = "bn254")]
 pub mod ntt;
+#[cfg(feature = "bn254")]
 pub mod optimize;
+#[cfg(feature = "bn254")]
 pub mod poseidon2;
 pub mod prefix_covector;
+#[cfg(feature = "bn254")]
 mod prover;
 mod r1cs;
+#[cfg(feature = "bn254")]
 pub mod skyscraper;
 pub mod sparse_matrix;
 mod transcript_sponge;
 pub mod u256_arith;
 pub mod utils;
+#[cfg(feature = "bn254")]
 mod verifier;
 mod whir_r1cs;
 pub mod witness;
@@ -23,18 +39,29 @@ use crate::{
     interner::{InternedFieldElement, Interner},
     sparse_matrix::{HydratedSparseMatrix, SparseMatrix},
 };
+/// The proof system's field. BN254 scalar field by default; the Goldilocks
+/// cubic extension (`Field64_3`, ~192 bits — `Field`/`FftField` but not
+/// `PrimeField`) under the `goldilocks` feature.
+#[cfg(feature = "bn254")]
+pub use ark_bn254::Fr as FieldElement;
+#[cfg(all(feature = "goldilocks", not(feature = "bn254")))]
+pub use whir::algebra::fields::Field64_3 as FieldElement;
+#[cfg(feature = "bn254")]
 pub use {
     acir::FieldElement as NoirElement,
-    ark_bn254::Fr as FieldElement,
-    hash_config::HashConfig,
     mavros::{MavrosProver, MavrosSchemeData},
     noir_proof_scheme::{NoirProof, NoirProofScheme, NoirSchemeData},
-    prefix_covector::{OffsetCovector, PrefixCovector, SparseCovector},
     prover::{NoirProver, Prover},
+    verifier::Verifier,
+};
+pub use {
+    hash_config::HashConfig,
+    prefix_covector::{OffsetCovector, PrefixCovector, SparseCovector},
     r1cs::R1CS,
     transcript_sponge::TranscriptSponge,
-    verifier::Verifier,
-    whir_r1cs::{R1csHash, WhirConfig, WhirR1CSProof, WhirR1CSScheme, WhirZkConfig},
+    whir_r1cs::{
+        R1csHash, WhirConfig, WhirR1CSProof, WhirR1CSScheme, WhirR1CSSchemeBuilder, WhirZkConfig,
+    },
     witness::PublicInputs,
 };
 
@@ -59,7 +86,11 @@ pub fn register_ntt() {
 
         // Register ProveKit-specific engines; WHIR's built-in engines
         // (SHA2, Keccak, Blake3, etc.) are pre-registered via whir::hash::ENGINES.
-        whir::hash::ENGINES.register(Arc::new(skyscraper::SkyscraperHashEngine));
-        whir::hash::ENGINES.register(Arc::new(poseidon2::Poseidon2HashEngine));
+        // Skyscraper and Poseidon2 are BN254-only constructions.
+        #[cfg(feature = "bn254")]
+        {
+            whir::hash::ENGINES.register(Arc::new(skyscraper::SkyscraperHashEngine));
+            whir::hash::ENGINES.register(Arc::new(poseidon2::Poseidon2HashEngine));
+        }
     });
 }
