+++ provekit/common/src/whir_r1cs.rs
@@ -1,12 +1,19 @@
+#[cfg(all(feature = "bn254", not(target_arch = "wasm32")))]
+use mavros_artifacts::R1CS as MavrosR1CS;
 #[cfg(debug_assertions)]
 use std::fmt::Debug;
 #[cfg(debug_assertions)]
 use whir::transcript::Interaction;
 use {
-    crate::{utils::serde_hex, FieldElement, HashConfig},
+    crate::{
+        utils::{next_power_of_two, serde_hex},
+        FieldElement, HashConfig, R1CS,
+    },
     serde::{Deserialize, Serialize},
     whir::{
         algebra::embedding::Identity,
+        engines::EngineId,
+        parameters::ProtocolParameters,
         protocols::{whir::Config as GenericWhirConfig, whir_zk::Config as GenericWhirZkConfig},
         transcript,
     },
@@ -74,6 +81,195 @@ impl WhirR1CSScheme {
     }
 }
 
+const MIN_WHIR_NUM_VARIABLES: usize = 13;
+const MIN_SUMCHECK_NUM_VARIABLES: usize = 1;
+
+/// Constructors for [`WhirR1CSScheme`], sizing the WHIR configuration from an
+/// R1CS instance (or raw dimensions) and stamping the instance-binding
+/// `r1cs_hash`.
+pub trait WhirR1CSSchemeBuilder {
+    /// Build a scheme for `r1cs`, stamping `r1cs_hash = r1cs.hash()`.
+    fn new_for_r1cs(
+        r1cs: &R1CS,
+        w1_size: usize,
+        num_challenges: usize,
+        challenge_offsets: Vec<usize>,
+        has_public_inputs: bool,
+        hash_config: HashConfig,
+    ) -> Self;
+
+    /// Build a scheme from a Mavros R1CS, leaving `r1cs_hash` unset.
+    #[cfg(all(feature = "bn254", not(target_arch = "wasm32")))]
+    fn new_from_mavros_r1cs(
+        r1cs: &MavrosR1CS,
+        w1_size: usize,
+        num_challenges: usize,
+        challenge_offsets: Vec<usize>,
+        has_public_inputs: bool,
+        hash_config: HashConfig,
+    ) -> Self;
+
+    /// Build a scheme from raw dimensions, leaving `r1cs_hash` unset.
+    fn new_from_dimensions(
+        num_witnesses: usize,
+        num_constraints: usize,
+        a_num_entries: usize,
+        w1_size: usize,
+        num_challenges: usize,
+        challenge_offsets: Vec<usize>,
+        has_public_inputs: bool,
+        hash_config: HashConfig,
+    ) -> Self;
+
+    /// Build the WHIR ZK configuration for `num_variables` (clamped to the
+    /// protocol minimum) and `num_polynomials`.
+    fn new_whir_zk_config_for_size(
+        num_variables: usize,
+        num_polynomials: usize,
+        hash_id: EngineId,
+    ) -> WhirZkConfig;
+}
+
+impl WhirR1CSSchemeBuilder for WhirR1CSScheme {
+    fn new_for_r1cs(
+        r1cs: &R1CS,
+        w1_size: usize,
+        num_challenges: usize,
+        challenge_offsets: Vec<usize>,
+        has_public_inputs: bool,
+        hash_config: HashConfig,
+    ) -> Self {
+        assert_eq!(
+            num_challenges,
+            challenge_offsets.len(),
+            "num_challenges ({num_challenges}) != challenge_offsets.len() ({})",
+            challenge_offsets.len()
+        );
+        let total_witnesses = r1cs.num_witnesses();
+        assert!(
+            w1_size <= total_witnesses,
+            "w1_size exceeds total witnesses"
+        );
+        let w2_size = total_witnesses - w1_size;
+
+        let m1_raw = next_power_of_two(w1_size);
+        let m2_raw = next_power_of_two(w2_size);
+        let m0_raw = next_power_of_two(r1cs.num_constraints());
+
+        let mut m_raw = m1_raw.max(m2_raw).max(MIN_WHIR_NUM_VARIABLES);
+        let m_0 = m0_raw.max(MIN_SUMCHECK_NUM_VARIABLES);
+
+        // Ensure w1's zero-padding has room for the blinding polynomial coefficients.
+        if (1usize << m_raw) - w1_size < 4 * m_0 {
+            m_raw += 1;
+        }
+
+        Self {
+            m: m_raw,
+            w1_size,
+            m_0,
+            a_num_terms: next_power_of_two(r1cs.a().iter().count()),
+            num_challenges,
+            challenge_offsets,
+            whir_witness: Self::new_whir_zk_config_for_size(m_raw, 1, hash_config.engine_id()),
+            has_public_inputs,
+            r1cs_hash: r1cs.hash(),
+            hash_config,
+        }
+    }
+
+    fn new_whir_zk_config_for_size(
+        num_variables: usize,
+        num_polynomials: usize,
+        hash_id: EngineId,
+    ) -> WhirZkConfig {
+        let nv = num_variables.max(MIN_WHIR_NUM_VARIABLES);
+
+        // Parameters tuned for 128-bit security under the Johnson bound (the old
+        // ConjectureList soundness was disproven). Rate=2 balances query count vs
+        // codeword size; ff=3 keeps blinding polynomials small; pow_bits=10 shifts
+        // security budget toward algebraic hardness (118 bits) with light PoW per
+        // round, which is faster than the default ~18-bit grinding.
+        let whir_params = ProtocolParameters {
+            unique_decoding: false,
+            security_level: 128,
+            pow_bits: 10,
+            initial_folding_factor: 3,
+            folding_factor: 3,
+            starting_log_inv_rate: 2,
+            batch_size: 1,
+            hash_id,
+        };
+        WhirZkConfig::new(1 << nv, &whir_params, num_polynomials)
+    }
+
+    #[cfg(all(feature = "bn254", not(target_arch = "wasm32")))]
+    fn new_from_mavros_r1cs(
+        r1cs: &MavrosR1CS,
+        w1_size: usize,
+        num_challenges: usize,
+        challenge_offsets: Vec<usize>,
+        has_public_inputs: bool,
+        hash_config: HashConfig,
+    ) -> Self {
+        let num_witnesses = r1cs.witness_layout.size();
+        let num_constraints = r1cs.constraints.len();
+        let a_num_entries: usize = r1cs.constraints.iter().map(|c| c.a.len()).sum();
+
+        Self::new_from_dimensions(
+            num_witnesses,
+            num_constraints,
+            a_num_entries,
+            w1_size,
+            num_challenges,
+            challenge_offsets,
+            has_public_inputs,
+            hash_config,
+        )
+    }
+
+    fn new_from_dimensions(
+        num_witnesses: usize,
+        num_constraints: usize,
+        a_num_entries: usize,
+        w1_size: usize,
+        num_challenges: usize,
+        challenge_offsets: Vec<usize>,
+        has_public_inputs: bool,
+        hash_config: HashConfig,
+    ) -> Self {
+        debug_assert_eq!(
+            num_challenges,
+            challenge_offsets.len(),
+            "num_challenges ({num_challenges}) != challenge_offsets.len() ({})",
+            challenge_offsets.len()
+        );
+        let m_raw = next_power_of_two(num_witnesses);
+        let m0_raw = next_power_of_two(num_constraints);
+
+        let mut m = m_raw.max(MIN_WHIR_NUM_VARIABLES);
+        let m_0 = m0_raw.max(MIN_SUMCHECK_NUM_VARIABLES);
+
+        // Ensure w1's zero-padding has room for the blinding polynomial coefficients.
+        if (1usize << m) - w1_size < 4 * m_0 {
+            m += 1;
+        }
+
+        Self {
+            m,
+            m_0,
+            a_num_terms: next_power_of_two(a_num_entries),
+            whir_witness: Self::new_whir_zk_config_for_size(m, 1, hash_config.engine_id()),
+            w1_size,
+            num_challenges,
+            challenge_offsets,
+            has_public_inputs,
+            r1cs_hash: R1csHash::UNSET,
+            hash_config,
+        }
+    }
+}
+
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct WhirR1CSProof {
     #[serde(with = "serde_hex")]
@@ -86,4 +282,53 @@ pub struct WhirR1CSProof {
     #[cfg(debug_assertions)]
     #[serde(skip)]
     pub pattern: Vec<Interaction>,
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn verify_security_level() {
+        let config = WhirR1CSScheme::new_whir_zk_config_for_size(20, 1, whir::hash::SHA2);
+        let sec_blinded = config
+            .blinded_commitment
+            .security_level(config.blinded_commitment.initial_committer.num_vectors, 1);
+        let sec_blinding = config
+            .blinding_commitment
+            .security_level(config.blinding_commitment.initial_committer.num_vectors, 1);
+        assert!(
+            sec_blinded >= 128.0,
+            "Blinded commitment security {sec_blinded:.2} < 128 bits"
+        );
+        assert!(
+            sec_blinding >= 128.0,
+            "Blinding commitment security {sec_blinding:.2} < 128 bits"
+        );
+    }
+
+    #[test]
+    fn verify_security_level_min_variables() {
+        let config = WhirR1CSScheme::new_whir_zk_config_for_size(
+            MIN_WHIR_NUM_VARIABLES,
+            1,
+            whir::hash::SHA2,
+        );
+        let sec_blinded = config
+            .blinded_commitment
+            .security_level(config.blinded_commitment.initial_committer.num_vectors, 1);
+        let sec_blinding = config
+            .blinding_commitment
+            .security_level(config.blinding_commitment.initial_committer.num_vectors, 1);
+        assert!(
+            sec_blinded >= 128.0,
+            "Blinded commitment security {sec_blinded:.2} < 128 bits at nv={}",
+            MIN_WHIR_NUM_VARIABLES
+        );
+        assert!(
+            sec_blinding >= 128.0,
+            "Blinding commitment security {sec_blinding:.2} < 128 bits at nv={}",
+            MIN_WHIR_NUM_VARIABLES
+        );
+    }
 }
