+++ provekit/r1cs-compiler/src/lib.rs
@@ -10,11 +10,10 @@ pub mod range_check;
 mod sha256_compression;
 mod spread;
 mod uints;
-mod whir_r1cs;
 mod witness_generator;
 
 pub use {
     noir_proof_scheme::{MavrosCompiler, NoirCompiler},
     noir_to_r1cs::{noir_to_r1cs, noir_to_r1cs_with_breakdown, R1CSBreakdown},
-    whir_r1cs::WhirR1CSSchemeBuilder,
+    provekit_common::WhirR1CSSchemeBuilder,
 };
